+++ CHANGELOG.md
@@ -128,6 +128,7 @@
   - `legacy_objc_type`
   - `legacy_random`
   - `lower_acl_than_parent`
+  - `multiline_arguments_brackets`
   - `multiline_parameters`
   - `multiple_closures_with_trailing_closure`
   - `no_extension_access_modifier`
@@ -321,6 +322,11 @@
   `trailing_newline` rule.  
   [SimplyDanny](https://github.com/SimplyDanny)
   [#4234](https://github.com/realm/SwiftLint/issues/4234)
+
+* Fix false-positives from `multiline_arguments_brackets` when a function call has a
+  single line trailing closure.  
+  [CraigSiemens](https://github.com/CraigSiemens)
+  [#4510](https://github.com/realm/SwiftLint/issues/4510)
 
 ## 0.49.1: Buanderie Principale